fix: make gRPC cert reset idempotent in current.ts (drop version-graph split)
Keep the cert reset in current.ts (in-place 26.6.1:2 bump) but make it
idempotent: only delete certs whose server.pem carries the legacy
`c-lightning.startos` SAN. That's safe to run on every update, so the
26.6.1:1 split (v26.6.1.1.ts + other: […]) isn't needed — reverted to
other: [].

The reset only matters for the 0.4.0-beta setupCerts era (25.12.1:x … 26.6:0);
native "cln" certs (0.3.5.1 and already-fixed installs) are left untouched.
Flagged (code TODO + TODO.md) for removal once 0.4.0 leaves beta.
